backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
konlpy==0.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2